Move Burn Header into v2 Burn folder
