+++ v2/components/Burn/package.json
@@ -5,11 +5,14 @@
   "version": "0.0.0",
   "dependencies": {
     "@chakra-ui/react": "^2.2.8",
+    "@snx-v2/CRatioHealthCard": "workspace:*",
+    "@snx-v2/Constants": "workspace:*",
     "@snx-v2/EthGasPriceEstimator": "workspace:*",
     "@snx-v2/ExternalLink": "workspace:*",
     "@snx-v2/MintOrBurnChanges": "workspace:*",
     "@snx-v2/TransactionModal": "workspace:*",
     "@snx-v2/formatters": "workspace:*",
+    "@snx-v2/getHealthVariant": "workspace:*",
     "@snx-v2/icons": "workspace:*",
     "@snx-v2/parseTxnError": "workspace:*",
     "@snx-v2/stakingCalculations": "workspace:*",
